pyproject.toml
# Please run `uv lock` to update uv.lock file if you change dependencies then run `uv sync` to install dependencies

[project]
name = "inviteease"
version = "0.1.0"
description = "InviteEase"

dependencies = [
    "Django==5.0.4",
    "django-import-export==4.3.7",
    "python-dotenv==1.1.0",
    "channels==4.1.0",
    "daphne==4.1.2",
    "whitenoise==6.6.0",
]
